nice.. altrep got cleaned
