+++ miniextendr-api/src/altrep_std_impls.rs
@@ -1,7 +1,10 @@
 //! Standard ALTREP backends (Vec/Arc/Slice/Mmap/Owned) split out for clarity.
 
-use crate::altrep::{IntBackend, LogicalBackend, RawBackend, RealBackend, StringBackend};
-use crate::ffi::{DATAPTR_RO, R_NaString, R_xlen_t, Rbyte, Rf_translateCharUTF8, SEXP, VECTOR_ELT};
+use crate::altrep::{IntBackend, ListBackend, LogicalBackend, RawBackend, RealBackend, StringBackend};
+use crate::ffi::{
+    DATAPTR_RO, LOGICAL_OR_NULL, R_NaString, R_xlen_t, Rbyte, Rf_translateCharUTF8, Rf_xlength,
+    SEXP, STRING_ELT, VECTOR_ELT,
+};
 use core::slice;
 use std::sync::{Arc, OnceLock};
 
@@ -215,7 +218,8 @@ impl Utf8Vec {
         let mut data = Vec::with_capacity(n as usize);
         let mut na = Vec::with_capacity(n as usize);
         for i in 0..n {
-            let ch = unsafe { VECTOR_ELT(x, i) };
+            // Use STRING_ELT for STRSXP (not VECTOR_ELT which is for VECSXP/lists)
+            let ch = unsafe { STRING_ELT(x, i) };
             if ch == unsafe { R_NaString } {
                 data.push(String::new());
                 na.push(true);
@@ -545,24 +549,42 @@ impl RawBackend for RawMmap {
     }
 }
 
-// Owned VECSXP helper
+/// Owned list of SEXP values for VECSXP.
+///
+/// # Warning
+/// This stores raw SEXP handles without R protection. The SEXPs must remain
+/// valid for the lifetime of this object. Typically used when the parent
+/// ALTREP object keeps the source protected.
 pub struct OwnedList {
     pub data: Vec<SEXP>,
 }
+// SAFETY: SEXP values are opaque pointers. The caller guarantees they remain valid.
+// GC safety is the caller's responsibility.
+unsafe impl Send for OwnedList {}
+unsafe impl Sync for OwnedList {}
 impl OwnedList {
     pub fn from_sexps(v: Vec<SEXP>) -> Self {
         Self { data: v }
     }
     /// # Safety
     /// `x` must be a VECSXP list. Elements are shallow-copied SEXP handles.
+    /// The caller must ensure the source remains protected from GC.
     pub unsafe fn from_list_sexp(x: SEXP) -> Self {
-        let n: R_xlen_t = unsafe { crate::ffi::Rf_xlength(x) };
+        let n: R_xlen_t = unsafe { Rf_xlength(x) };
         let mut v = Vec::with_capacity(n as usize);
         for i in 0..n {
             let elt = unsafe { VECTOR_ELT(x, i) };
             v.push(elt);
         }
         Self { data: v }
+    }
+}
+impl ListBackend for OwnedList {
+    fn len(&self) -> R_xlen_t {
+        self.data.len() as R_xlen_t
+    }
+    fn elt(&self, i: R_xlen_t) -> SEXP {
+        self.data[i as usize]
     }
 }
 
@@ -597,5 +619,182 @@ impl IntBackend for IntSliceMat {
             .materialized
             .get_or_init(|| self.src.to_vec().into_boxed_slice());
         Some(&**bx)
+    }
+}
+
+// Int Vec/Arc/Mmap backends
+pub struct IntVec {
+    pub data: Box<[i32]>,
+}
+impl From<Vec<i32>> for IntVec {
+    fn from(v: Vec<i32>) -> Self {
+        Self {
+            data: v.into_boxed_slice(),
+        }
+    }
+}
+impl IntBackend for IntVec {
+    fn len(&self) -> R_xlen_t {
+        self.data.len() as R_xlen_t
+    }
+    fn elt(&self, i: R_xlen_t) -> i32 {
+        self.data[i as usize]
+    }
+    fn get_region(&self, i: R_xlen_t, n: R_xlen_t, out: &mut [i32]) -> R_xlen_t {
+        let end = (i + n).min(self.len()) as usize;
+        let src = &self.data[i as usize..end];
+        out[..src.len()].copy_from_slice(src);
+        src.len() as R_xlen_t
+    }
+    fn dataptr(&self) -> Option<&[i32]> {
+        Some(&self.data)
+    }
+}
+
+pub struct IntArc {
+    pub data: Arc<[i32]>,
+}
+impl From<Arc<[i32]>> for IntArc {
+    fn from(data: Arc<[i32]>) -> Self {
+        Self { data }
+    }
+}
+impl IntBackend for IntArc {
+    fn len(&self) -> R_xlen_t {
+        self.data.len() as R_xlen_t
+    }
+    fn elt(&self, i: R_xlen_t) -> i32 {
+        self.data[i as usize]
+    }
+    fn get_region(&self, i: R_xlen_t, n: R_xlen_t, out: &mut [i32]) -> R_xlen_t {
+        let end = (i + n).min(self.len()) as usize;
+        let src = &self.data[i as usize..end];
+        out[..src.len()].copy_from_slice(src);
+        src.len() as R_xlen_t
+    }
+    fn dataptr(&self) -> Option<&[i32]> {
+        Some(&self.data)
+    }
+}
+
+pub struct IntMmap {
+    pub ptr: *const i32,
+    pub len: usize,
+    pub cleanup: Option<unsafe extern "C-unwind" fn(*const i32, usize)>,
+}
+// SAFETY: IntMmap's raw pointer points to externally-managed memory (e.g. mmap).
+// The caller guarantees the pointer remains valid for the object's lifetime.
+unsafe impl Send for IntMmap {}
+unsafe impl Sync for IntMmap {}
+impl IntMmap {
+    /// # Safety
+    /// Caller guarantees the pointer and length designate a valid, readable buffer
+    /// for the lifetime of the ALTREP object or until `cleanup` is called.
+    pub unsafe fn new(
+        ptr: *const i32,
+        len: usize,
+        cleanup: Option<unsafe extern "C-unwind" fn(*const i32, usize)>,
+    ) -> Self {
+        Self { ptr, len, cleanup }
+    }
+}
+impl Drop for IntMmap {
+    fn drop(&mut self) {
+        if let Some(f) = self.cleanup {
+            unsafe { f(self.ptr, self.len) }
+        }
+    }
+}
+impl IntBackend for IntMmap {
+    fn len(&self) -> R_xlen_t {
+        self.len as R_xlen_t
+    }
+    fn elt(&self, i: R_xlen_t) -> i32 {
+        unsafe { *self.ptr.add(i as usize) }
+    }
+    fn get_region(&self, i: R_xlen_t, n: R_xlen_t, out: &mut [i32]) -> R_xlen_t {
+        let start = i as usize;
+        let end = ((i + n).min(self.len())) as usize;
+        let src = unsafe { slice::from_raw_parts(self.ptr.add(start), end - start) };
+        out[..src.len()].copy_from_slice(src);
+        src.len() as R_xlen_t
+    }
+    fn dataptr(&self) -> Option<&[i32]> {
+        Some(unsafe { slice::from_raw_parts(self.ptr, self.len) })
+    }
+}
+
+/// Owned contiguous i32 buffer for LOGICALSXP.
+pub struct OwnedLogical {
+    pub data: Box<[i32]>,
+}
+impl OwnedLogical {
+    /// # Safety
+    /// `x` must be a LGLSXP; caller guarantees it is valid and points to readable memory.
+    pub unsafe fn from_lgls_sexp(x: SEXP) -> Self {
+        unsafe {
+            let n = Rf_xlength(x) as usize;
+            let ptr = LOGICAL_OR_NULL(x);
+            let data = if ptr.is_null() {
+                Vec::new().into_boxed_slice()
+            } else {
+                slice::from_raw_parts(ptr, n).to_vec().into_boxed_slice()
+            };
+            Self { data }
+        }
+    }
+}
+impl LogicalBackend for OwnedLogical {
+    fn len(&self) -> R_xlen_t {
+        self.data.len() as R_xlen_t
+    }
+    fn elt(&self, i: R_xlen_t) -> i32 {
+        self.data[i as usize]
+    }
+    fn get_region(&self, i: R_xlen_t, n: R_xlen_t, out: &mut [i32]) -> R_xlen_t {
+        let end = (i + n).min(self.len()) as usize;
+        let src = &self.data[i as usize..end];
+        out[..src.len()].copy_from_slice(src);
+        src.len() as R_xlen_t
+    }
+    fn dataptr(&self) -> Option<&[i32]> {
+        Some(&self.data)
+    }
+    fn no_na(&self) -> i32 {
+        0
+    }
+}
+
+/// Owned contiguous raw buffer for RAWSXP.
+pub struct OwnedRaw {
+    pub data: Box<[Rbyte]>,
+}
+impl OwnedRaw {
+    /// # Safety
+    /// `x` must be a RAWSXP; caller guarantees it is valid and points to readable memory.
+    pub unsafe fn from_raw_sexp(x: SEXP) -> Self {
+        unsafe {
+            let n = Rf_xlength(x) as usize;
+            let ptr = DATAPTR_RO(x) as *const Rbyte;
+            let data = slice::from_raw_parts(ptr, n).to_vec().into_boxed_slice();
+            Self { data }
+        }
+    }
+}
+impl RawBackend for OwnedRaw {
+    fn len(&self) -> R_xlen_t {
+        self.data.len() as R_xlen_t
+    }
+    fn elt(&self, i: R_xlen_t) -> Rbyte {
+        self.data[i as usize]
+    }
+    fn get_region(&self, i: R_xlen_t, n: R_xlen_t, out: &mut [Rbyte]) -> R_xlen_t {
+        let end = (i + n).min(self.len()) as usize;
+        let src = &self.data[i as usize..end];
+        out[..src.len()].copy_from_slice(src);
+        src.len() as R_xlen_t
+    }
+    fn dataptr(&self) -> Option<&[Rbyte]> {
+        Some(&self.data)
     }
 }
